+++ NEWS.rst
@@ -7,6 +7,12 @@ Removals
 ------------------------------
 * `(defn/a …)` is now `(defn :async …)`.
 * `(fn/a …)` is now `(fn :async …)`.
+* `(with/a […] …)` is now `(with [:async …] …)`.
+
+   * As with `for`, `:async` must precede each name to be bound
+     asynchronously, because you can mix synchronous and asynchronous
+     types.
+
 * `(yield-from …)` is now `(yield :from …)`.
 
 New Features